Bugfix for edgecase in sum_waveform
Errors appeared in a rare situation where the zero-padded part of a record overlaps with a peak, causing a nonphysical selection range. 'Lucky'  enough this situation was not so rare in XAMS-type data :) I believe the right behavior is a ''continue'' meaning to skip this record in the peak consideration.