New print_entry function
This one is optional (and I don't mind discarting this change if not needed): it is to print an entry of an np.ndarray in a human-readable format. I have found I need it when reviewing data. If there is anything like pd.DataFrame.head() for numpy arrays, //please// let me know.