Make exception work for incomplete matadata too
This is a great feature of strax that is needed if one e.g. wants to load only a small part of the data. The exception is also gread in case the RunMetaData is not available. 

However in case of incomplete RunMetaData the error is not a strax.RunMetadataNotAvailable but a KeyError since 'start' is not in the RunMetaData. Hence I'd propose to also accept the KeyError as an exception.

I've tested in on eb0 for run 7157 and 7158 while loading raw_records with incomplete RunMetaData.