Correctly check endtimes in chunk.py (#281)
We observed that although the data is always sorted by (start)time, this is not per se true for the endtime. Therefore the last record may not have the latest endtime. This is the safest way to get around this.

Depending on the datarate one might also argue that for example we look at the last 500 samples :
```
data_ends_at = strax.endtime(self.data[-500:]).max()
```
* Check last 500 samples for data_ends_at